Feat - Integrate date picker for expiration date
Added the `@react-native-community/datetimepicker` dependency and integrated it into the `AddDocumentDetailsSheet` component. This change replaced the manual text input for the expiration date with a native date picker, improving user experience and input accuracy by providing a more intuitive way to select dates. The date picker is displayed conditionally based on user interaction.

Co-authored-by: Diego Gaspa <diego.gaspariseo@udlap.mx>