Feat - Add option to hide tags on FolderCard
Added a `showTags` prop to the `FolderCard` component, allowing its tags section to be conditionally rendered. This new capability was utilized within the `FolderModal` to hide tags on the folder type examples, thereby streamlining the interface. Concurrently, the breadcrumb navigation text in `FolderHeader` was localized to "Regresar a", and minor adjustments were made to its padding and margin for improved layout.